graph: Move to wibox.widget
